% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_storr.R
\name{map_from_storr}
\alias{map_from_storr}
\title{Map a function over a list of storr objects}
\usage{
map_from_storr(input_keys, storr, .f, ...,
  .input_namespace = storr$default_namespace, .flush_cache = FALSE)
}
\arguments{
\item{input_keys}{List of keys present in \code{storr}}

\item{storr}{Storr object}

\item{.f}{Function to run on each object}

\item{...}{Other arguments passed to \code{.f}}

\item{.input_namespace}{Namespace of \code{input_keys} (optional)}

\item{.flush_cache}{Flush cache of objects at each step? Defaults to \code{TRUE}}
}
\value{
A list.
}
\description{
This variant saves the \link[purrr:map]{purrr::map} results within the current environment.
}
