% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_storr.R
\name{map_within_storr}
\alias{map_within_storr}
\title{Map inputs from storr into outputs also saved in storr}
\usage{
map_within_storr()
}
\description{
Map inputs from storr into outputs also saved in storr
}
