% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_storr.R
\name{map_from_storr}
\alias{map_from_storr}
\title{Map a function over a list of storr objects}
\usage{
map_from_storr(keys, storr, namespace = storr$default_namespace, .f, ...,
  .keep_cache = FALSE)
}
\arguments{
\item{keys}{List of keys present in `storr`}

\item{storr}{Storr object}

\item{namespace}{Namespace of keys (optional)}

\item{.f}{Function to run on each object retrieved from `keys`}

\item{...}{Other arguments passed to `.f`}

\item{.keep_cache}{Keep the retrieved objects cached? Defaults to `FALSE`}
}
\value{
A list resulting from map.
}
\description{
This variant saves the [purrr::map] results within the current environment.
}
